backend/requirements.txt
schedule
selenium
webdriver_manager
python-telegram-bot